Make i_model implementation responsible for reporting readiness for scoring
The model_downloader class was treating "empty data" as an invalid model_data to pass to i_model implementations. For pdf_model, there is no model_data, so we never got to model_data_received.

This change switches the notion to "model readiness", rather than a flag depending on presense/absence of data - and changes the api for update() to report out whether the model is ready to score.