Implement PDF Passthrough (Sample Mode) (#41)
# Implement "PDF passthrough" model type

This uses a singleton instance of safe_vw to manage the vw state, but instantiates it without a model. This requires a manual call to refresh_model, and setting no background_updates.

* Add a test for pdf_model, verifying the PDF
* Expose refresh_model in python API
* Enable use of PdfPassthrough model in C# Adds ability to get and set string values in configuration
* Enables manual refresh for model

# Enable "no model data" (e.g. PDF input) scenarios

The default Azure Blob model source does not work when there is not a semantic meaning to getting a model blob - it forces an annoying workaround of pointing to a valid HTTP GET-able resource.

The fix is to enable a first-class "No Model Data" scenario.

* Add a "No Model Data" as a model "source"
* Change pdf-passthrough test to use new "no model data" mechanism
* Spelling: restapi_data_tranport => restapi_data_transport

# Make i_model implementation responsible for reporting readiness for scoring

The model_downloader class was treating "empty data" as an invalid model_data to pass to i_model implementations. For pdf_model, there is no model_data, so we never got to model_data_received.

This change switches the notion to "model readiness", rather than a flag depending on presense/absence of data - and changes the api for update() to report out whether the model is ready to score.

# Refactor i_model API surface to output pdf

Moves the sampling code out of i_model implementations and into live_model. This prevents implementors of i_model to have to worry about sampling, and ensures it is consistent between different underlying models.

## Other fixes

* Fix Configuration[key].get to be consistent with default value.
* Fix spelling
    "Unkown" => "Unknown"
    "recieved" => "received"