Fix memory corruption in during ISender creation
In certain timing situations, it is possible that the sender_create_fn delegate gets GCed before it has a chance to be used. It needs to be valid for the lifetime of FactoryContext.

The fix is to introduce a GCHandle holding on to the delegate. We wrap it inside of GCHandleLifetime to make it easier to clean up on multiple invocations to Set