Introduce dedup+zstd suppport.
This is done by introducing i_logger_extensions which hides all dedup behavior and state from the rest of the codebase.
The main touchpoint is in logger_facade.cc, that uses it to produce transformed events.
The use of the is_enabled() call is to shortcut to a fast path when available.

Finally, this commit includes test_tools/log_parser/parser.py which can parse rlclientlib output when using the file
backend.