Support ONNX Models via OnnxRuntime on Linux (#110)
Adds support for CB-like ONNX models using OnnxRuntime on Linux

"CB-like" is defined as having at least one tensor output that will be interpreted as a PDF for the purposes of sampling and assigning a "chosen action".

The Context is provided through a JSON-like "TensorNotation". The format is:

```
Example:

{"input1":"BAAAAAAAAAA=;AACAP2ZmBkBmZoZAmpkRwQ=="}

This represents a rank-1 tensor "input1" of shape { 4 } with values { 1.0f, 2.1f, 4.2f, -9.1f }

Details:

<TENSORS> := "{" <TENSOR-LIST> "}"
<TENSOR-LIST> := <TENSOR> ["," [<TENSOR-LIST>]]
<TENSOR> := "\"" <INPUT-NAME> "\":\"" <TENSOR-DATA> "\""
<TENSOR-DATA> := <DIMS-BASE64> ";" <VALUES-BASE64>
<DIMS-BASE64> := { base64 encoding of int64[] representing dimensions of the tensor }
<VALUES-BASE64> := { base64 encoding of float[] representing values of the tensor }
```

No other JSON concepts are allowed. The reason for any relation to JSON is the current semantics of RLClientLib and associated tools expecting to log the context as JSON. Note that whitespace is not allowed.