Schema v2 rework schema to make events payloads completely free-standing.
One main realization of when we tried to use the new schema is that
systems would have to carry a lot of side information to payloads in
order to produce joined logs.

This PR addresses this by moving all data required to parse an event
into the event payload itself. Namely:

- Batch content type now only informs of whether dedup was used or not.
Events carry themselves whether they are compressed or not.

-EventBatch now holds an array of SerializedEvent, which is just a container
to a serialized Event object.

Beyod that, it's now posible to enable/disable dedup and compression separatedly
by using interaction.send.use_compression and interaction.send.use_dedup config variables.