Fix a bunch of issues found by VS 2019 compiled and  /permissive-
In particular:

Make sure `container_iterator` and `const_container_iterator` correctly
work as random access iterators.

Fix `container_iterator::operator*()const` to return a non-const value.
A const interator is not the same thing as a const_interator. C++
constness is a constant source of joyfull surprises.

Beyond that, fix a couple variable size issues and a missing return code
in a falible function.