Pass i_trace all the way to async_batcher so it can log stuff. Add a dtor to live_model_impl and flush the queues from it.
Why flushing from live_model_impl is ok but from async_batcher is not?

async_batcher holds un-owned references to a lot of stuff and the
expectation is that those objects won't be valid during destruction.

Flushing data in async_batcher requires those dangling pointers to
be valid so we cannot do it from the dtor.

Then why is it ok to do it from live_model_impl dtor? As it turns out,
it holds owning references to pretty much everything that's needed for flushing.

Did you say pretty much? Yes, there are a lot of un-owned references in
live_model_impl as well but they look related to initialization only.
Life is too short otherwise.