fix: make launch_vs command set envvars correctly (#335)
The previous version of this was setting the BaseOutputFolder (which is changed under CMake when building DotNet bindings to ensure that native dlls are next to the managed dlls that require them) with quotes inside the variable. The fix is to quote the full set arg, so that quotes are not included in the value.