[binary-log] Add CLI override to checkpoint parameters. (#360)
CLI overrides must be sticky to be useful, so we introduce the notion of
sticky values that once set, can't be modified.

This is done by having a sticky_value<T> container in which the setter
takes an argument which controls the stickyness of the value. This way
we can use it for both CLI sources and file sources.