fix: Set and check input_serialization settings correctly
There is a fair bit of logic in RLClientLib which is very payload-specific, particularly around dedup and the more complex loop types.

In the initial support for FlatBuffer contexts we are going to avoid enabling those scenarios to get to E2E FlatBuffer CB, so we suppress these capabilities when we use FlatBuffer.

The fix is to do this correctly and in all cases so we do not run into weird AccessViolation exceptions / SegFaults.

The longer-term fix would be to provide a higher-level way to interact with the context (likely via the underlying i_model) to avoid having to deeply understand every single input_serialization type. This is a fairly significant refactor, and beyond the scope of the current effort around FlatBuffers.